Stub validated commit v2
