+++ xmtp_mls/src/groups/group_permissions.rs
@@ -17,8 +17,9 @@ use xmtp_proto::xmtp::mls::message_contents::{
 };
 
 use super::{
-    group_mutable_metadata::{GroupMutableMetadata, GroupMutableMetadataError},
-    validated_commit::{AggregatedMembershipChange, CommitParticipant, ValidatedCommit},
+    group_mutable_metadata::GroupMutableMetadata,
+    // validated_commit::{AggregatedMembershipChange, CommitParticipant, ValidatedCommit},
+    validated_commit::{CommitParticipant, Inbox, MetadataChange, ValidatedCommit},
 };
 
 // A trait for policies that can update Metadata for the group
@@ -98,24 +99,6 @@ impl MetadataPolicies {
 
     pub fn any(policies: Vec<MetadataPolicies>) -> Self {
         MetadataPolicies::AnyCondition(MetadataAnyCondition::new(policies))
-    }
-}
-
-// Information for Metadata Update used for validation
-#[derive(Clone, Debug)]
-pub struct MetadataChange {
-    pub(crate) old_value: GroupMutableMetadata,
-    pub(crate) new_value: GroupMutableMetadata,
-    pub(crate) metadata_policies: HashMap<String, MetadataPolicies>,
-}
-
-impl Default for MetadataChange {
-    fn default() -> Self {
-        Self {
-            old_value: GroupMutableMetadata::default(),
-            new_value: GroupMutableMetadata::default(),
-            metadata_policies: MetadataPolicies::default_map(MetadataPolicies::allow()),
-        }
     }
 }
 
@@ -249,7 +232,7 @@ impl MetadataPolicy for MetadataAnyCondition {
 
 // A trait for policies that can add/remove members and installations for the group
 pub trait MembershipPolicy: std::fmt::Debug {
-    fn evaluate(&self, actor: &CommitParticipant, change: &AggregatedMembershipChange) -> bool;
+    fn evaluate(&self, actor: &CommitParticipant, change: &Inbox) -> bool;
     fn to_proto(&self) -> Result<MembershipPolicyProto, PolicyError>;
 }
 
@@ -277,11 +260,11 @@ pub enum BasePolicies {
 }
 
 impl MembershipPolicy for BasePolicies {
-    fn evaluate(&self, actor: &CommitParticipant, change: &AggregatedMembershipChange) -> bool {
+    fn evaluate(&self, actor: &CommitParticipant, inbox: &Inbox) -> bool {
         match self {
             BasePolicies::Allow => true,
             BasePolicies::Deny => false,
-            BasePolicies::AllowSameMember => change.account_address == actor.account_address,
+            BasePolicies::AllowSameMember => actor.inbox_id == inbox.inbox_id,
             BasePolicies::AllowIfActorCreator => actor.is_creator,
         }
     }
@@ -378,11 +361,11 @@ impl TryFrom<MembershipPolicyProto> for MembershipPolicies {
 }
 
 impl MembershipPolicy for MembershipPolicies {
-    fn evaluate(&self, actor: &CommitParticipant, change: &AggregatedMembershipChange) -> bool {
+    fn evaluate(&self, actor: &CommitParticipant, inbox: &Inbox) -> bool {
         match self {
-            MembershipPolicies::Standard(policy) => policy.evaluate(actor, change),
-            MembershipPolicies::AndCondition(policy) => policy.evaluate(actor, change),
-            MembershipPolicies::AnyCondition(policy) => policy.evaluate(actor, change),
+            MembershipPolicies::Standard(policy) => policy.evaluate(actor, inbox),
+            MembershipPolicies::AndCondition(policy) => policy.evaluate(actor, inbox),
+            MembershipPolicies::AnyCondition(policy) => policy.evaluate(actor, inbox),
         }
     }
 
@@ -408,10 +391,10 @@ impl AndCondition {
 }
 
 impl MembershipPolicy for AndCondition {
-    fn evaluate(&self, actor: &CommitParticipant, change: &AggregatedMembershipChange) -> bool {
+    fn evaluate(&self, actor: &CommitParticipant, inbox: &Inbox) -> bool {
         self.policies
             .iter()
-            .all(|policy| policy.evaluate(actor, change))
+            .all(|policy| policy.evaluate(actor, inbox))
     }
 
     fn to_proto(&self) -> Result<MembershipPolicyProto, PolicyError> {
@@ -441,10 +424,10 @@ impl AnyCondition {
 }
 
 impl MembershipPolicy for AnyCondition {
-    fn evaluate(&self, actor: &CommitParticipant, change: &AggregatedMembershipChange) -> bool {
+    fn evaluate(&self, actor: &CommitParticipant, inbox: &Inbox) -> bool {
         self.policies
             .iter()
-            .any(|policy| policy.evaluate(actor, change))
+            .any(|policy| policy.evaluate(actor, inbox))
     }
 
     fn to_proto(&self) -> Result<MembershipPolicyProto, PolicyError> {
@@ -465,33 +448,9 @@ impl MembershipPolicy for AnyCondition {
 pub struct PolicySet {
     pub add_member_policy: MembershipPolicies,
     pub remove_member_policy: MembershipPolicies,
-    pub add_installation_policy: MembershipPolicies,
-    pub remove_installation_policy: MembershipPolicies,
     pub update_metadata_policy: HashMap<String, MetadataPolicies>,
 }
 
-fn extract_field_changed(change: &MetadataChange) -> Result<String, GroupMutableMetadataError> {
-    let changes: Vec<&String> = change
-        .old_value
-        .attributes
-        .iter()
-        .filter(|(key, old_value)| {
-            match change.new_value.attributes.get(*key) {
-                Some(new_value) => &new_value != old_value,
-                None => true, // Assuming missing keys in `new_value` count as changes
-            }
-        })
-        .map(|(key, _)| key)
-        .collect();
-
-    match changes.len() {
-        1 => Ok(changes[0].clone()), // There is exactly one change
-        0 => Err(GroupMutableMetadataError::NoUpdates),
-        _ => Err(GroupMutableMetadataError::TooManyUpdates),
-    }
-}
-
-#[allow(dead_code)]
 impl PolicySet {
     pub fn new(
         add_member_policy: MembershipPolicies,
@@ -501,30 +460,24 @@ impl PolicySet {
         Self {
             add_member_policy,
             remove_member_policy,
-            add_installation_policy: default_add_installation_policy(),
-            remove_installation_policy: default_remove_installation_policy(),
             update_metadata_policy,
         }
     }
 
     pub fn evaluate_commit(&self, commit: &ValidatedCommit) -> bool {
         self.evaluate_policy(
-            commit.members_added.iter(),
+            commit.added_inboxes.iter(),
             &self.add_member_policy,
             &commit.actor,
         ) && self.evaluate_policy(
-            commit.members_removed.iter(),
+            commit.removed_inboxes.iter(),
             &self.remove_member_policy,
             &commit.actor,
-        ) && self.evaluate_policy(
-            commit.installations_added.iter(),
-            &self.add_installation_policy,
+        ) && self.evaluate_metadata_policy(
+            commit.metadata_changes.iter(),
+            &self.update_metadata_policy,
             &commit.actor,
-        ) && self.evaluate_policy(
-            commit.installations_removed.iter(),
-            &self.remove_installation_policy,
-            &commit.actor,
-        ) & self.evaluate_metadata_policy(&commit.group_name_updated, &commit.actor)
+        )
     }
 
     fn evaluate_policy<'a, I, P>(
@@ -534,7 +487,7 @@ impl PolicySet {
         actor: &CommitParticipant,
     ) -> bool
     where
-        I: Iterator<Item = &'a AggregatedMembershipChange>,
+        I: Iterator<Item = &'a Inbox>,
         P: MembershipPolicy + std::fmt::Debug,
     {
         changes.all(|change| {
@@ -551,44 +504,35 @@ impl PolicySet {
         })
     }
 
-    // In case group creator is on future version of libxmtp, we can validate
-    // metadata policies on new unknown fields
-    fn evaluate_metadata_policy(&self, change: &MetadataChange, actor: &CommitParticipant) -> bool {
-        #[allow(clippy::needless_late_init)]
-        let field_changed;
-        match extract_field_changed(change) {
-            Ok(f) => field_changed = f,
-            Err(error) => {
-                match error {
-                    // If there is no change in metadata, no need to validate the policy
-                    GroupMutableMetadataError::NoUpdates => return true,
-                    _ => {
-                        log::info!(
-                            "Change extraction failed for actor {:?} and change {:?}",
-                            actor,
-                            change
-                        );
-                        return false;
-                    }
+    fn evaluate_metadata_policy<'a, I>(
+        &self,
+        mut changes: I,
+        policies: &HashMap<String, MetadataPolicies>,
+        actor: &CommitParticipant,
+    ) -> bool
+    where
+        I: Iterator<Item = &'a MetadataChange>,
+    {
+        changes.all(|change| {
+            if let Some(policy) = policies.get(&change.field_name) {
+                let is_ok = policy.evaluate(actor, change);
+                if !is_ok {
+                    log::info!(
+                        "Policy for field {} failed for actor {:?} and change {:?}",
+                        change.field_name,
+                        actor,
+                        change
+                    );
+                    return false;
                 }
+                return is_ok;
             }
-        }
-
-        if let Some(policy) = change.metadata_policies.get(&field_changed) {
-            let is_ok = policy.evaluate(actor, change);
-            if !is_ok {
-                log::info!(
-                    "Policy {:?} failed for actor {:?} and change {:?}",
-                    policy,
-                    actor,
-                    change
-                );
-            }
-            is_ok
-        } else {
-            log::info!("Missing policy for the changed field: {:?}", &field_changed);
+            log::info!(
+                "Missing policy for changed metadata field: {}",
+                change.field_name
+            );
             false
-        }
+        })
     }
 
     pub(crate) fn to_proto(&self) -> Result<PolicySetProto, PolicyError> {
@@ -642,20 +586,9 @@ impl PolicySet {
     }
 }
 
-fn default_add_installation_policy() -> MembershipPolicies {
-    MembershipPolicies::allow()
-}
-
-fn default_remove_installation_policy() -> MembershipPolicies {
-    MembershipPolicies::deny()
-}
-
 /// A policy where any member can add or remove any other member
 pub(crate) fn policy_everyone_is_admin() -> PolicySet {
-    let mut metadata_policies_map: HashMap<String, MetadataPolicies> = HashMap::new();
-    for field in GroupMutableMetadata::supported_fields() {
-        metadata_policies_map.insert(field.to_string(), MetadataPolicies::allow());
-    }
+    let metadata_policies_map = MetadataPolicies::default_map(MetadataPolicies::allow());
     PolicySet::new(
         MembershipPolicies::allow(),
         MembershipPolicies::allow(),
@@ -665,13 +598,8 @@ pub(crate) fn policy_everyone_is_admin() -> PolicySet {
 
 /// A policy where only the group creator can add or remove members
 pub(crate) fn policy_group_creator_is_admin() -> PolicySet {
-    let mut metadata_policies_map: HashMap<String, MetadataPolicies> = HashMap::new();
-    for field in GroupMutableMetadata::supported_fields() {
-        metadata_policies_map.insert(
-            field.to_string(),
-            MetadataPolicies::allow_if_actor_creator(),
-        );
-    }
+    let metadata_policies_map =
+        MetadataPolicies::default_map(MetadataPolicies::allow_if_actor_creator());
     PolicySet::new(
         MembershipPolicies::allow_if_actor_creator(),
         MembershipPolicies::allow_if_actor_creator(),
@@ -713,29 +641,27 @@ impl std::fmt::Display for PreconfiguredPolicies {
 
 #[cfg(test)]
 mod tests {
-    use crate::utils::test::{rand_account_address, rand_vec};
+    use crate::{
+        groups::group_mutable_metadata::MetadataField,
+        utils::test::{rand_string, rand_vec},
+    };
 
     use super::*;
 
-    fn build_change(
-        account_address: Option<String>,
-        installation_id: Option<Vec<u8>>,
-        is_creator: bool,
-    ) -> AggregatedMembershipChange {
-        AggregatedMembershipChange {
-            account_address: account_address.unwrap_or_else(rand_account_address),
-            installation_ids: vec![installation_id.unwrap_or_else(rand_vec)],
+    fn build_change(inbox_id: Option<String>, is_creator: bool) -> Inbox {
+        Inbox {
+            inbox_id: inbox_id.unwrap_or(rand_string()),
             is_creator,
         }
     }
 
     fn build_actor(
-        account_address: Option<String>,
+        inbox_id: Option<String>,
         installation_id: Option<Vec<u8>>,
         is_creator: bool,
     ) -> CommitParticipant {
         CommitParticipant {
-            account_address: account_address.unwrap_or_else(rand_account_address),
+            inbox_id: inbox_id.unwrap_or(rand_string()),
             installation_id: installation_id.unwrap_or_else(rand_vec),
             is_creator,
         }
@@ -745,38 +671,31 @@ mod tests {
         // Add a member with the same account address as the actor if true, random account address if false
         member_added: Option<bool>,
         member_removed: Option<bool>,
-        installation_added: Option<bool>,
-        installation_removed: Option<bool>,
+        metadata_fields_changed: Option<Vec<String>>,
         actor_is_creator: bool,
     ) -> ValidatedCommit {
         let actor = build_actor(None, None, actor_is_creator);
         let build_membership_change = |same_address_as_actor| {
             if same_address_as_actor {
-                vec![build_change(
-                    Some(actor.account_address.clone()),
-                    None,
-                    actor_is_creator,
-                )]
+                vec![build_change(Some(actor.inbox_id.clone()), actor_is_creator)]
             } else {
-                vec![build_change(None, None, false)]
+                vec![build_change(None, false)]
             }
         };
 
         ValidatedCommit {
             actor: actor.clone(),
-            members_added: member_added
+            added_inboxes: member_added
                 .map(build_membership_change)
                 .unwrap_or_default(),
-            members_removed: member_removed
+            removed_inboxes: member_removed
                 .map(build_membership_change)
                 .unwrap_or_default(),
-            installations_added: installation_added
-                .map(build_membership_change)
-                .unwrap_or_default(),
-            installations_removed: installation_removed
-                .map(build_membership_change)
-                .unwrap_or_default(),
-            group_name_updated: MetadataChange::default(),
+            metadata_changes: metadata_fields_changed
+                .unwrap_or(vec![])
+                .into_iter()
+                .map(|field| MetadataChange::new(field, Some(rand_string()), Some(rand_string())))
+                .collect(),
         }
     }
 
@@ -790,7 +709,7 @@ mod tests {
             MetadataPolicies::default_map(MetadataPolicies::allow()),
         );
 
-        let commit = build_validated_commit(Some(true), Some(true), None, None, false);
+        let commit = build_validated_commit(Some(true), Some(true), None, false);
         assert!(permissions.evaluate_commit(&commit));
     }
 
@@ -802,21 +721,11 @@ mod tests {
             MetadataPolicies::default_map(MetadataPolicies::deny()),
         );
 
-        let member_added_commit = build_validated_commit(Some(false), None, None, None, false);
+        let member_added_commit = build_validated_commit(Some(false), None, None, false);
         assert!(!permissions.evaluate_commit(&member_added_commit));
 
-        let member_removed_commit = build_validated_commit(None, Some(false), None, None, false);
+        let member_removed_commit = build_validated_commit(None, Some(false), None, false);
         assert!(!permissions.evaluate_commit(&member_removed_commit));
-
-        let installation_added_commit =
-            build_validated_commit(None, None, Some(false), None, false);
-        // Installation added is always allowed
-        assert!(permissions.evaluate_commit(&installation_added_commit));
-
-        // Installation removed is always denied
-        let installation_removed_commit =
-            build_validated_commit(None, None, None, Some(false), false);
-        assert!(!permissions.evaluate_commit(&installation_removed_commit));
     }
 
     #[test]
@@ -827,11 +736,10 @@ mod tests {
             MetadataPolicies::default_map(MetadataPolicies::deny()),
         );
 
-        let commit_with_creator = build_validated_commit(Some(true), Some(true), None, None, true);
+        let commit_with_creator = build_validated_commit(Some(true), Some(true), None, true);
         assert!(permissions.evaluate_commit(&commit_with_creator));
 
-        let commit_without_creator =
-            build_validated_commit(Some(true), Some(true), None, None, false);
+        let commit_without_creator = build_validated_commit(Some(true), Some(true), None, false);
         assert!(!permissions.evaluate_commit(&commit_without_creator));
     }
 
@@ -843,11 +751,10 @@ mod tests {
             MetadataPolicies::default_map(MetadataPolicies::deny()),
         );
 
-        let commit_with_same_member = build_validated_commit(Some(true), None, None, None, false);
+        let commit_with_same_member = build_validated_commit(Some(true), None, None, false);
         assert!(permissions.evaluate_commit(&commit_with_same_member));
 
-        let commit_with_different_member =
-            build_validated_commit(Some(false), None, None, None, false);
+        let commit_with_different_member = build_validated_commit(Some(false), None, None, false);
         assert!(!permissions.evaluate_commit(&commit_with_different_member));
     }
 
@@ -862,7 +769,7 @@ mod tests {
             MetadataPolicies::default_map(MetadataPolicies::deny()),
         );
 
-        let member_added_commit = build_validated_commit(Some(true), None, None, None, false);
+        let member_added_commit = build_validated_commit(Some(true), None, None, false);
         assert!(!permissions.evaluate_commit(&member_added_commit));
     }
 
@@ -877,7 +784,7 @@ mod tests {
             MetadataPolicies::default_map(MetadataPolicies::deny()),
         );
 
-        let member_added_commit = build_validated_commit(Some(true), None, None, None, false);
+        let member_added_commit = build_validated_commit(Some(true), None, None, false);
         assert!(permissions.evaluate_commit(&member_added_commit));
     }
 
@@ -903,6 +810,32 @@ mod tests {
         let restored = PolicySet::from_bytes(as_bytes.as_slice()).expect("proto conversion failed");
         // All fields implement PartialEq so this should test equality all the way down
         assert!(permissions.eq(&restored))
+    }
+
+    #[test]
+    fn test_update_group_name() {
+        let allow_permissions = PolicySet::new(
+            MembershipPolicies::allow(),
+            MembershipPolicies::allow(),
+            MetadataPolicies::default_map(MetadataPolicies::allow()),
+        );
+
+        let member_added_commit = build_validated_commit(
+            Some(true),
+            None,
+            Some(vec![MetadataField::GroupName.to_string()]),
+            false,
+        );
+
+        assert!(allow_permissions.evaluate_commit(&member_added_commit));
+
+        let deny_permissions = PolicySet::new(
+            MembershipPolicies::allow(),
+            MembershipPolicies::allow(),
+            MetadataPolicies::default_map(MetadataPolicies::deny()),
+        );
+
+        assert!(!deny_permissions.evaluate_commit(&member_added_commit));
     }
 
     #[test]
