+++ xmtp_mls/src/retry.rs
@@ -111,7 +111,7 @@ impl Retry {
 /// # Example
 ///  ```
 ///  use thiserror::Error;
-///  use xmtp_mls::{retry, retry::{RetryableError, Retry}};
+///  use xmtp_mls::{retry_sync, retry::{RetryableError, Retry}};
 ///
 /// #[derive(Debug, Error)]
 /// enum MyError {
@@ -140,7 +140,7 @@ impl Retry {
 ///
 /// fn main() {
 ///      let mut i = 0;
-///      retry!(Retry::default(), (|| -> Result<(), MyError> {
+///      retry_sync!(Retry::default(), (|| -> Result<(), MyError> {
 ///         let res = fallable_fn(i);
 ///         i += 1;
 ///         res
@@ -149,7 +149,7 @@ impl Retry {
 ///  }
 /// ```
 #[macro_export]
-macro_rules! retry {
+macro_rules! retry_sync {
     ($retry: expr, $code: tt) => {{
         #[allow(unused)]
         use $crate::retry::RetryableError;
@@ -302,7 +302,7 @@ mod tests {
             Ok(())
         };
 
-        retry!(Retry::default(), test_fn).unwrap();
+        retry_sync!(Retry::default(), test_fn).unwrap();
     }
 
     #[test]
@@ -317,7 +317,7 @@ mod tests {
             retryable_with_args(i, "Hello".to_string(), &list)
         };
 
-        retry!(Retry::default(), test_fn).unwrap();
+        retry_sync!(Retry::default(), test_fn).unwrap();
     }
 
     #[test]
@@ -326,7 +326,7 @@ mod tests {
             retry_error_fn()?;
             Ok(())
         };
-        let result: Result<(), SomeError> = retry!(Retry::default(), (closure));
+        let result: Result<(), SomeError> = retry_sync!(Retry::default(), (closure));
 
         assert!(result.is_err())
     }
@@ -339,7 +339,7 @@ mod tests {
             Err(SomeError::DontRetryThis)
         };
 
-        let _r = retry!(Retry::default(), test_fn);
+        let _r = retry_sync!(Retry::default(), test_fn);
 
         assert_eq!(attempts, 1);
     }
